@fuhrysteve 2017-07-26
Looks like there's a depreciation warning in there that suggests using `hash_password` instead - does it make sense to change this PR to use `hash_password` instead of `encrypt_password`?