@sh4r3m4n 2017-07-27
Hi @fuhrysteve, you're right, I changed this to use hash_password instead. Thanks for the advice!